# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST de Gestión de Vehículos",
        "description": "API RESTful para gestión de vehículos, marcas y propietarios desarrollada con Laravel",
        "contact": {
            "name": "Equipo de Desarrollo",
            "email": "contacto@vehiculos-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
...
        "/api/vehiculos": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Obtener lista de todos los vehículos",
                "description": "Retorna una lista completa de todos los vehículos con sus marcas y propietarios",
                "operationId": "5e171b6b3e06799480137d4ce4b5bced",
                "responses": {
                    "200": {
                        "description": "Lista de vehículos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vehículos obtenidos exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vehicle"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Crear un nuevo vehículo",
                "description": "Registra un nuevo vehículo en el sistema",
                "operationId": "ef9122ef6c2925f07878b02fe59b2968",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "model",
                                    "vehicle_brand_id",
                                    "number_of_doors",
                                    "color"
                                ],
                                "properties": {
                                    "model": {
                                        "description": "Modelo del vehículo",
                                        "type": "string",
                                        "example": "Corolla 2024"
                                    },
                                    "vehicle_brand_id": {
                                        "description": "ID de la marca del vehículo",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "number_of_doors": {
                                        "description": "Número de puertas (mínimo 2, máximo 6)",
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "color": {
                                        "description": "Color del vehículo",
                                        "type": "string",
                                        "example": "Blanco"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehículo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vehículo creado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vehiculos/{id}": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Obtener un vehículo específico",
                "description": "Retorna los detalles de un vehículo por su ID, incluyendo marca y propietarios",
                "operationId": "e5e0e02b37494424c834cc1f1b8dc481",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehículo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vehículo obtenido exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Actualizar un vehículo existente",
                "description": "Actualiza los datos de un vehículo",
                "operationId": "c5d7449d371bc99973e8756a1ec11c89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "model",
                                    "vehicle_brand_id",
                                    "number_of_doors",
                                    "color"
                                ],
                                "properties": {
                                    "model": {
                                        "type": "string",
                                        "example": "Camry 2025"
                                    },
                                    "vehicle_brand_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "number_of_doors": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Negro"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehículo actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vehículo actualizado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Eliminar un vehículo",
                "description": "Elimina un vehículo del sistema (también elimina sus asociaciones con propietarios)",
                "operationId": "158ba982bc96048e59b93e5da89e6c90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehículo eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vehículo eliminado exitosamente"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vehiculos/{id}/propietarios": {
            "post": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Asignar un propietario a un vehículo",
                "description": "Asocia una persona como propietaria de un vehículo (relación many-to-many)",
                "operationId": "390eb1dd8c3960de9c0ee43e33c4acfa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "person_id"
                                ],
                                "properties": {
                                    "person_id": {
                                        "description": "ID de la persona a asignar como propietaria",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Propietario asignado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Propietario asignado exitosamente al vehículo"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "La persona ya es propietaria del vehículo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error en la operación"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error de validación"
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "El campo es obligatorio."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Person": {
                "title": "Persona",
                "description": "Modelo de persona/propietario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "identification_number": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Vehicle": {
                "title": "Vehículo",
                "description": "Modelo de vehículo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "model": {
                        "type": "string",
                        "example": "Corolla 2024"
                    },
                    "vehicle_brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "number_of_doors": {
                        "type": "integer",
                        "example": 4
                    },
                    "color": {
                        "type": "string",
                        "example": "Blanco"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "vehicle_brand": {
                        "$ref": "#/components/schemas/VehicleBrand"
                    },
                    "owners": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Person"
                        }
                    }
                },
                "type": "object"
            },
            "VehicleBrand": {
                "title": "Marca de Vehículo",
                "description": "Modelo de marca de vehículo",
                "properties": {
                    "id": {
                        "description": "ID de la marca",
                        "type": "integer",
                        "example": 1
                    },
                    "brand_name": {
                        "description": "Nombre de la marca",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "country": {
                        "description": "País de origen",
                        "type": "string",
                        "example": "Japón"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-10 10:30:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-10 10:30:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}